Move local state query helpers to separate module
