+++ lib/network-layer/cardano-wallet-network-layer.cabal
@@ -47,6 +47,7 @@ library
     Cardano.Wallet.Network.Implementation.Ouroboros
     Cardano.Wallet.Network.Implementation.UnliftIO
     Cardano.Wallet.Network.Light
+    Cardano.Wallet.Network.LocalStateQuery.Extra
     Cardano.Wallet.Network.Logging
     Cardano.Wallet.Network.Logging.Aggregation
